fix: changed the conversion to take a float as arg

+++ util/conversion.go
@@ -1,6 +1,6 @@
 package util
 
-func BytesToGB(bytes int) float64 {
+func BytesToGB(bytes float64) float64 {
 	gigabytes := float64(bytes) / (1e+9)
 	return gigabytes
 }
